fix: Composite pattern implementation
