+++ src/AMQPConnection.php
@@ -9,15 +9,14 @@ use PhpAmqpLib\Channel\AMQPChannel;
 use PhpAmqpLib\Connection\AMQPSSLConnection;
 use PhpAmqpLib\Connection\AMQPStreamConnection;
 
-class AMQPConnection
+class AMQPConnection extends AMQPConnectionType
 {
     private AMQPStreamConnection $connection;
     private AMQPChannel $channel;
-    private array $config;
 
     public function __construct()
     {
-        $this->config = config('amqp');
+        parent::__construct(config('amqp'));
         $this->setConnection();
     }
 
@@ -69,10 +68,11 @@ class AMQPConnection
     {
         try {
             if (!is_null($this->config)) {
-                Arr::get($this->config, 'ssl', true)
-                    ? $this->setSSLConnection()
-                    : $this->setStreamConnection();
-
+                if (Arr::get($this->config, 'ssl', true)) {
+                    $this->connection = parent::getSSLConnection();
+                } else {
+                    $this->connection = parent::getStreamConnection();
+                }
                 $this->connection->set_close_on_destruct(true);
                 $this->channel = $this->connection->channel();
             } else {
@@ -81,54 +81,5 @@ class AMQPConnection
         } catch (Exception $ex) {
             throw new AMQPConnectionException("Error ->" . $ex->getMessage());
         }
-    }
-
-    /**
-     * Set a standard stream connection with config file info.
-     * 
-     * @return void
-     */
-    private function setStreamConnection(): void
-    {
-        $host = $this->getHostConfig();
-        $this->connection = new AMQPStreamConnection(
-            Arr::get($host, 'host'),
-            Arr::get($host, 'port'),
-            Arr::get($host, 'username'),
-            Arr::get($host, 'password'),
-            Arr::get($host, 'vhost')
-        );
-    }
-
-    /**
-     * Set a SSL connection with config file info.
-     * 
-     * @return void
-     */
-    private function setSSLConnection(): void
-    {
-        $sslOptions = array_filter(Arr::get($this->config, 'ssl_options', []), function ($item) {
-            return null !== $item;
-        });
-
-        $host = $this->getHostConfig();
-        $this->connection = new AMQPSSLConnection(
-            Arr::get($host, 'host'),
-            Arr::get($host, 'port'),
-            Arr::get($host, 'user'),
-            Arr::get($host, 'password'),
-            Arr::get($host, 'vhost'),
-            $sslOptions
-        );
-    }
-
-    /**
-     * Get host info in the config file
-     * 
-     * @return array with host info
-     */
-    private function getHostConfig(): array
-    {
-        return Arr::get($this->config, 'host');
     }
 }
